+++ source/library/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_subdirectory(foundation)
-add_subdirectory(core)
+add_subdirectory(runtime)
 add_subdirectory(filesystem)
 add_subdirectory(gpu)
 add_subdirectory(render)
