+++ source/library/concurrency/public/potato/concurrency/concurrent_queue.h
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#include <potato/core/assertion.h>
+#include <potato/runtime/assertion.h>
 #include "potato/foundation/box.h"
 #include "potato/foundation/int_types.h"
 #include <atomic>
