+++ source/library/foundation/public/potato/foundation/delegate.h
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#include <potato/core/assertion.h>
+#include <potato/runtime/assertion.h>
 #include "traits.h"
 
 #include <new>
