+++ source/library/gpu/private/d3d11_backend/d3d11_command_list.cpp
@@ -8,7 +8,7 @@
 #include "d3d11_texture.h"
 #include "d3d11_sampler.h"
 #include "potato/foundation/int_types.h"
-#include <potato/core/assertion.h>
+#include <potato/runtime/assertion.h>
 #include "potato/foundation/out_ptr.h"
 
 up::d3d11::CommandListD3D11::CommandListD3D11(com_ptr<ID3D11DeviceContext> context) : _context(std::move(context)) {}
