+++ source/library/gpu/private/d3d11_backend/d3d11_texture.cpp
@@ -4,7 +4,7 @@
 #include "d3d11_platform.h"
 #include "potato/gpu/com_ptr.h"
 #include "potato/foundation/out_ptr.h"
-#include <potato/core/assertion.h>
+#include <potato/runtime/assertion.h>
 
 up::d3d11::TextureD3D11::TextureD3D11(com_ptr<ID3D11Resource> texture) : _texture(std::move(texture)) {
 }
