+++ source/library/runtime/private/CMakeLists.txt
@@ -1,9 +1,9 @@
-target_sources(core PRIVATE
+target_sources(runtime PRIVATE
     $<$<NOT:$<PLATFORM_ID:Windows>>:${CMAKE_CURRENT_SOURCE_DIR}/callstack.posix.cpp>
     $<$<PLATFORM_ID:Windows>:${CMAKE_CURRENT_SOURCE_DIR}/callstack.windows.cpp>
 )
 
-target_sources(core PRIVATE
+target_sources(runtime PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/debug.cpp
     $<$<NOT:$<PLATFORM_ID:Windows>>:${CMAKE_CURRENT_SOURCE_DIR}/debug.posix.cpp>
     $<$<PLATFORM_ID:Windows>:${CMAKE_CURRENT_SOURCE_DIR}/debug.windows.cpp>
