+++ source/library/runtime/public/potato/runtime/CMakeLists.txt
@@ -1,4 +1,4 @@
-target_sources(core PRIVATE
+target_sources(runtime PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/_export.h
     ${CMAKE_CURRENT_SOURCE_DIR}/assertion.h
     ${CMAKE_CURRENT_SOURCE_DIR}/callstack.h
