+++ source/library/runtime/public/potato/runtime/debug.h
@@ -30,7 +30,7 @@ namespace up {
     };
 
     /// Ask the user if they should abort, skip the error, or always skip this error.
-    UP_CORE_API UP_NOINLINE error_action fatal_error(char const* file, int line, char const* failedConditionText, char const* messageText);
+    UP_RUNTIME_API UP_NOINLINE error_action fatal_error(char const* file, int line, char const* failedConditionText, char const* messageText);
 } // namespace up
 
 #define _up_FAIL(failure, message) \
