+++ source/library/runtime/tests/test_callstack.cpp
@@ -1,4 +1,4 @@
-#include "potato/core/callstack.h"
+#include "potato/runtime/callstack.h"
 #include <doctest/doctest.h>
 #include <array>
